# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Live Test",
        "description": "This is api testing",
        "contact": {
            "email": "xyz@xyz.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/price_comparision/{type}": {
            "get": {
                "tags": [
                    "Price Comparision"
                ],
                "summary": "Price Comparision",
                "operationId": "price-comparision",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/price_comparision/store": {
            "post": {
                "tags": [
                    "Price Comparision"
                ],
                "summary": "Store Price Comparision",
                "operationId": "store-price-comparision",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/price_comparision/details": {
            "post": {
                "tags": [
                    "Price Comparision"
                ],
                "summary": "Send Price Comparision",
                "operationId": "send-price-comparision",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}